# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "free-ask-bot swagger server.",
        "title": "free-ask-bot API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1.0",
    "paths": {
        "/clusters": {
            "get": {
                "description": "分页获取集群列表",
                "tags": [
                    "Cluster"
                ],
                "summary": "集群列表",
                "parameters": [
                    {
                        "type": "integer",
                        "x-example": "1",
                        "description": "分页号，默认为1",
                        "name": "pageNo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "x-example": "10",
                        "description": "分页大小，默认为10",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查询条件，默认为空",
                        "name": "filterKey",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询值，默认为空",
                        "name": "filterValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应",
                        "schema": {
                            "$ref": "#/definitions/model.ListClusterResp"
                        }
                    }
                }
            },
            "post": {
                "description": "根据配置参数自动创建集群",
                "tags": [
                    "Cluster"
                ],
                "summary": "创建集群",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "CreateClusterReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateClusterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应",
                        "schema": {
                            "$ref": "#/definitions/model.CreateClusterResp"
                        }
                    }
                }
            }
        },
        "/clusters/{clusterId}": {
            "get": {
                "description": "根据集群ID获取集群详细信息",
                "tags": [
                    "Cluster"
                ],
                "summary": "集群详情",
                "parameters": [
                    {
                        "type": "string",
                        "x-example": "cluster-sedqqz7ka",
                        "description": "集群资源ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应",
                        "schema": {
                            "$ref": "#/definitions/model.GetClusterResp"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据集群ID自动删除集群",
                "tags": [
                    "Cluster"
                ],
                "summary": "删除集群",
                "parameters": [
                    {
                        "type": "string",
                        "x-example": "cluster-sedqqz7ka",
                        "description": "集群资源ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应",
                        "schema": {
                            "$ref": "#/definitions/model.BaseResponse"
                        }
                    }
                }
            }
        },
        "/tools/check-cidr": {
            "post": {
                "description": "检查集群CIDR是否存在网段冲突",
                "tags": [
                    "Tools"
                ],
                "summary": "CIDR网段检查",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "CheckCIDRReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CheckCIDRReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应",
                        "schema": {
                            "$ref": "#/definitions/model.CheckCIDRResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "Success"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "调用成功"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string",
                    "example": "6893b1e9-da8f-4c6c-a161-eba4b81ea5b3"
                }
            }
        },
        "model.CheckCIDRReq": {
            "type": "object",
            "required": [
                "podCidr",
                "serviceCidr"
            ],
            "properties": {
                "ipv6": {
                    "description": "ipv6开关",
                    "type": "boolean",
                    "example": false
                },
                "podCidr": {
                    "description": "容器组网段",
                    "type": "string",
                    "example": "10.244.0.0/16"
                },
                "podCidrIpv6": {
                    "description": "ipv6容器组网段, ipv6为true时有效",
                    "type": "string",
                    "example": "fc00::/48"
                },
                "serviceCidr": {
                    "description": "服务网段",
                    "type": "string",
                    "example": "10.96.0.0/16"
                },
                "serviceCidrIpv6": {
                    "description": "ipv6服务网",
                    "type": "string",
                    "example": "fd00::/108"
                }
            }
        },
        "model.CheckCIDRResp": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "Success"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "调用成功"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string",
                    "example": "6893b1e9-da8f-4c6c-a161-eba4b81ea5b3"
                }
            }
        },
        "model.ClusterDetail": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "amazing-cluster-description"
                },
                "k8sType": {
                    "description": "类型：k8s、k3s",
                    "type": "string",
                    "example": "k8s"
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "example": "imortal-cluster-name"
                },
                "resourceID": {
                    "description": "集群ID",
                    "type": "string",
                    "example": "cluster-sedqqz7ka"
                },
                "runtime": {
                    "description": "容器运行时",
                    "type": "string",
                    "example": "cri-o"
                },
                "status": {
                    "description": "集群状态",
                    "type": "string"
                },
                "version": {
                    "description": "版本",
                    "type": "string",
                    "example": "1.22.5"
                }
            }
        },
        "model.ClusterSummary": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "amazing-cluster-description"
                },
                "k8sType": {
                    "description": "类型：k8s、k3s",
                    "type": "string",
                    "example": "k8s"
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "example": "imortal-cluster-name"
                },
                "resourceID": {
                    "description": "集群ID",
                    "type": "string",
                    "example": "cluster-sedqqz7ka"
                },
                "runtime": {
                    "description": "容器运行时",
                    "type": "string",
                    "example": "cri-o"
                },
                "status": {
                    "description": "集群状态",
                    "type": "string"
                },
                "version": {
                    "description": "版本",
                    "type": "string",
                    "example": "1.22.5"
                }
            }
        },
        "model.CreateClusterReq": {
            "type": "object",
            "required": [
                "k8sType",
                "runtime",
                "version"
            ],
            "properties": {
                "description": {
                    "description": "描述，支持 0~255 位字符",
                    "type": "string",
                    "example": "amazing-cluster-description"
                },
                "k8sType": {
                    "description": "类型：k8s、k3s",
                    "type": "string",
                    "enum": [
                        "k8s",
                        "k3s"
                    ],
                    "example": "k8s"
                },
                "name": {
                    "description": "名称，支持 1～127 位字符，必须以字母或中文开头，可以包含字母、数字、下划线（_）、中划线（-）、点(.)",
                    "type": "string",
                    "example": "imortal-cluster-name"
                },
                "runtime": {
                    "description": "容器运行时",
                    "type": "string",
                    "enum": [
                        "cri-o",
                        "containerd"
                    ],
                    "example": "cri-o"
                },
                "version": {
                    "description": "版本",
                    "type": "string",
                    "example": "1.22.5"
                }
            }
        },
        "model.CreateClusterResp": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "Success"
                },
                "data": {
                    "description": "集群资源ID",
                    "type": "string",
                    "example": "cluster-sedqqz7kavbh"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "调用成功"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string",
                    "example": "6893b1e9-da8f-4c6c-a161-eba4b81ea5b3"
                }
            }
        },
        "model.GetClusterResp": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "Success"
                },
                "data": {
                    "description": "集群详情",
                    "$ref": "#/definitions/model.ClusterDetail"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "调用成功"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string",
                    "example": "6893b1e9-da8f-4c6c-a161-eba4b81ea5b3"
                }
            }
        },
        "model.ListClusterResp": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "Success"
                },
                "data": {
                    "description": "集群概要列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ClusterSummary"
                    }
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "调用成功"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string",
                    "example": "6893b1e9-da8f-4c6c-a161-eba4b81ea5b3"
                },
                "totalCount": {
                    "description": "集群总数",
                    "type": "integer",
                    "example": 100
                }
            }
        }
    }
}